backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-dotenv
google-generativeai
pydantic
watchfiles
requests
